Move statistic rendering into component
