+++ src/parser/core/modules/Statistics/Statistics.tsx
@@ -1,18 +1,11 @@
-import classNames from 'classnames'
 import Module, {DISPLAY_MODE} from 'parser/core/Module'
 import React from 'react'
-import {Icon} from 'semantic-ui-react'
 import DISPLAY_ORDER from '../DISPLAY_ORDER'
+import {StatisticComponent} from './StatisticComponent'
 import styles from './Statistics.module.css'
 
 // tslint:disable-next-line:no-magic-numbers
-type ColumnSpan = 1 | 2 | 3 | 4
-const spanClassMap: Record<ColumnSpan, keyof typeof styles> = {
-	1: 'span1',
-	2: 'span2',
-	3: 'span3',
-	4: 'span4',
-}
+export type ColumnSpan = 1 | 2 | 3 | 4
 
 export interface Statistic {
 	Content: React.ComponentType
@@ -41,39 +34,9 @@ export class Statistics extends Module {
 	output() {
 		return (
 			<div className={styles.statistics}>
-				{testStatistics.map((statistic, i) => {
-					const colSpanClass = spanClassMap[statistic.width || 1]
-					return (
-						<div
-							key={i}
-							className={classNames(
-								styles.statistic,
-								styles[colSpanClass],
-								statistic.Info && styles.hasInfo,
-							)}
-							style={{gridRowEnd: `span ${statistic.height || 1}`}}
-						>
-							<div className={styles.content}>
-								<statistic.Content/>
-							</div>
-
-							{statistic.Info && (
-								<div className={styles.info}>
-									<Icon name="info" className={styles.infoIcon}/>
-								</div>
-							)}
-						</div>
-					)
-				})}
-
-				<div className= {classNames(styles.statistic, styles.span2)}/>
-				<div className={styles.statistic}/>
-				<div className={classNames(styles.statistic, styles.span2)}/>
-				<div className={styles.statistic} style={{gridRowEnd: 'span 2'}}/>
-				<div className={styles.statistic}/>
-				<div className={styles.statistic}/>
-				<div className={classNames(styles.statistic, styles.span3)}/>
-				<div className={styles.statistic}/>
+				{testStatistics.map((statistic, i) => (
+					<StatisticComponent key={i} statistic={statistic}/>
+				))}
 			</div>
 		)
 	}
